CountryItem changes

+++ app/src/main/java/com/example/worlddata/CountryItem.kt
@@ -7,5 +7,5 @@ package com.example.worlddata
 data class CountryItem(
 	val name: String,
 	val flag: Int,
-	val parameter: Long
+	val parameter: Double // todo: it should be a String.
 )
